Migrate the Qemu controller to a custom controller example
This controller contains additional dependencies we haven't fully
handled, hasn't been thoroughly tested, and is even unfinished,
but contains valuable code that could be an example for how to
derive a custom controller backend.
