+++ guibot/inputmap.py
@@ -438,96 +438,6 @@ class VNCDoToolKey(Key):
         self.KP_DECIMAL = None
 
 
-class QemuKey(Key):
-    """Helper to contain all key mappings for the Qemu DC backend."""
-
-    def __init__(self):
-        """Build an instance containing the key map for the Qemu backend."""
-        super().__init__()
-
-        self.ENTER = 'ret'
-        self.TAB = 'tab'
-        self.ESC = 'esc'
-        self.BACKSPACE = 'backspace'
-        self.DELETE = 'delete'
-        self.INSERT = 'insert'
-
-        self.CTRL = 'ctrl'
-        self.ALT = 'alt'
-        # TODO: if needed these are also supported
-        # altgr, altgr_r (right altgr)
-        self.SHIFT = 'shift'
-        # TODO: 'meta' is not available
-        self.META = None
-        self.RCTRL = 'ctrl_r'
-        self.RALT = 'alt_r'
-        self.RSHIFT = 'shift_r'
-        # TODO: 'right meta' is not available
-        self.RMETA = None
-
-        self.F1 = 'f1'
-        self.F2 = 'f2'
-        self.F3 = 'f3'
-        self.F4 = 'f4'
-        self.F5 = 'f5'
-        self.F6 = 'f6'
-        self.F7 = 'f7'
-        self.F8 = 'f8'
-        self.F9 = 'f9'
-        self.F10 = 'f10'
-        self.F11 = 'f11'
-        self.F12 = 'f12'
-        # TODO: these function keys are not available
-        self.F13 = None
-        self.F14 = None
-        self.F15 = None
-        self.F16 = None
-        self.F17 = None
-        self.F18 = None
-        self.F19 = None
-        self.F20 = None
-
-        self.HOME = 'home'
-        self.END = 'end'
-        self.LEFT = 'left'
-        self.RIGHT = 'right'
-        self.UP = 'up'
-        self.DOWN = 'down'
-        self.PAGE_DOWN = 'pgdn'
-        self.PAGE_UP = 'pgup'
-
-        self.CAPS_LOCK = 'caps_lock'
-        self.PRINTSCREEN = 'print'
-        # TODO: 'pause' is not available
-        self.PAUSE = None
-        self.SCROLL_LOCK = 'scroll_lock'
-        self.NUM_LOCK = 'num_lock'
-        self.SYS_REQ = 'sysrq'
-        self.SUPER = '0xdc'
-        self.RSUPER = '0xdb'
-        # TODO: 'hyper' and 'right hyper' are not available
-        self.HYPER = None
-        self.RHYPER = None
-        self.MENU = "menu"
-
-        self.KP0 = 'kp_0'
-        self.KP1 = 'kp_1'
-        self.KP2 = 'kp_2'
-        self.KP3 = 'kp_3'
-        self.KP4 = 'kp_4'
-        self.KP5 = 'kp_5'
-        self.KP6 = 'kp_6'
-        self.KP7 = 'kp_7'
-        self.KP8 = 'kp_8'
-        self.KP9 = 'kp_9'
-        self.KP_ENTER = 'kp_enter'
-        self.KP_DIVIDE = 'kp_divide'
-        self.KP_MULTIPLY = 'kp_multiply'
-        self.KP_SUBTRACT = 'kp_subtract'
-        self.KP_ADD = 'kp_add'
-        self.KP_DECIMAL = 'kp_decimal'
-
-
 class PyAutoGUIKey(Key):
     """Helper to contain all key mappings for the PyAutoGUI DC backend."""
 
@@ -694,22 +604,6 @@ class VNCDoToolKeyModifier(KeyModifier):
         self.MOD_META = 'meta'
 
 
-class QemuKeyModifier(KeyModifier):
-    """Helper to contain all modifier key mappings for the Qemu DC backend."""
-
-    def __init__(self):
-        """Build an instance containing the modifier key map for the Qemu backend."""
-        super().__init__()
-
-        # TODO: 'none' is not available
-        self.MOD_NONE = None
-        self.MOD_CTRL = 'ctrl'
-        self.MOD_ALT = 'alt'
-        self.MOD_SHIFT = 'shift'
-        # TODO: 'meta' is not available
-        self.MOD_META = None
-
-
 class PyAutoGUIKeyModifier(KeyModifier):
     """Helper to contain all modifier key mappings for the PyAutoGUI DC backend."""
 
@@ -787,18 +681,6 @@ class VNCDoToolMouseButton(MouseButton):
 
         self.LEFT_BUTTON = 1
         self.RIGHT_BUTTON = 3
-        self.CENTER_BUTTON = 2
-
-
-class QemuMouseButton(MouseButton):
-    """Helper to contain all mouse button mappings for the Qemu DC backend."""
-
-    def __init__(self):
-        """Build an instance containing the mouse button map for the Qemu backend."""
-        super().__init__()
-
-        self.LEFT_BUTTON = 1
-        self.RIGHT_BUTTON = 4
         self.CENTER_BUTTON = 2
